cast Statfs_t.Frsize to int64 for s390x arch

+++ driver/node.go
@@ -453,9 +453,9 @@ func (s *nodeService) NodeGetVolumeStats(ctx context.Context, req *csi.NodeGetVo
 	if sfs.Blocks > 0 {
 		usage = append(usage, &csi.VolumeUsage{
 			Unit:      csi.VolumeUsage_BYTES,
-			Total:     int64(sfs.Blocks) * sfs.Frsize,
-			Used:      int64(sfs.Blocks-sfs.Bfree) * sfs.Frsize,
-			Available: int64(sfs.Bavail) * sfs.Frsize,
+			Total:     int64(sfs.Blocks) * int64(sfs.Frsize),
+			Used:      int64(sfs.Blocks-sfs.Bfree) * int64(sfs.Frsize),
+			Available: int64(sfs.Bavail) * int64(sfs.Frsize),
 		})
 	}
 	if sfs.Files > 0 {
